Rename generate search index script
