+++ gulp/school-search-index.js
@@ -1,5 +1,5 @@
 const gulp = require('gulp')
-const lunr = require('./../scripts/generate_search_index') // generate
+const lunr = require('./../scripts/generate-search-index') // generate
 
 gulp.task('school-search-index', (done) => {
   // Generate search index
