+++ scripts/generate-search-index.js
